fraud-detection-win-arBenim/fraud-detection-win-ar/requirements.txt
pandas==2.0.3
numpy>=1.23,<2.0
scikit-learn==1.3.2
matplotlib
seaborn
xgboost==2.0.3
imbalanced-learn==0.11.0
PyYAML
joblib
